refactor(carousel): refactor stories
affects: @atlantis-lab/carousel

- Merged two stories into a single one
