+++ package.json
@@ -9,17 +9,10 @@
     "typecheck": "mctl typecheck",
     "checkcommit": "mctl precommit",
     "postinstall": "lerna run compile",
-    "build": "lerna run build",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "build": "lerna run build"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.10",
     "@monstrs/mctl": "0.1.82",
-    "@storybook/addon-actions": "^6.1.15",
-    "@storybook/addon-essentials": "^6.1.15",
-    "@storybook/addon-links": "^6.1.15",
-    "@storybook/react": "^6.1.15",
     "@types/react": "^17.0.0",
     "babel-loader": "^8.2.2",
     "husky": "4.3.0",
