+++ packages/carousel/.storybook/main.js
@@ -1,0 +1,4 @@
+module.exports = {
+  stories: ['../src/**/*.story.mdx', '../src/**/*.story.@(js|jsx|ts|tsx)'],
+  addons: ['@storybook/addon-links', '@storybook/addon-essentials'],
+}
